requirements.txt
loguru
web3
aiohttp
ton-client-py
mypy
ruff
